refactor(components.ng-ui-router-breadcrumb): renamed package name
replace package name from `@ovh-ux/ng-uirouter-breadcrumb` to `@ovh-ux/ng-ui-router-breadcrumb`
and also renamed module name from `ngUirouterBreadcrumb` to `ngUiRouterBreadcrumb`

BREAKING CHANGE: package and module are now respectively named as
`@ovh-ux/ng-ui-router-breadcrumb` and `ngUiRouterBreadcrumb`
