+++ README.md
@@ -62,8 +62,8 @@ yarn install
 
 | Package | Version | Dependencies | Dev dependencies | Peer dependencies | Changelog |
 | ------- | ------- | ------------ | ---------------- | ----------------- | :-------: |
+| [@ovh-ux/ng-ui-router-breadcrumb](https://github.com/ovh-ux/manager/tree/master/packages/components/ng-ui-router-breadcrumb) | [![npm version](https://badgen.net/npm/v/@ovh-ux/ng-ui-router-breadcrumb)](https://www.npmjs.com/package/@ovh-ux/ng-ui-router-breadcrumb) | [![Dependencies](https://badgen.net/david/dep/ovh-ux/manager/packages/components/ng-ui-router-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-ui-router-breadcrumb?activeTab=dependencies) | [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/manager/packages/components/ng-ui-router-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-ui-router-breadcrumb?activeTab=dependencies) | [![Peer Dependencies](https://badgen.net/david/peer/ovh-ux/manager/packages/components/ng-ui-router-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-ui-router-breadcrumb?activeTab=dependencies) | [:books:](https://github.com/ovh-ux/manager/blob/master/packages/components/ng-ui-router-breadcrumb/CHANGELOG.md) |
 | [@ovh-ux/ng-ui-router-layout](https://github.com/ovh-ux/manager/tree/master/packages/components/ng-ui-router-layout) | [![npm version](https://badgen.net/npm/v/@ovh-ux/ng-ui-router-layout)](https://www.npmjs.com/package/@ovh-ux/ng-ui-router-layout) | [![Dependencies](https://badgen.net/david/dep/ovh-ux/manager/packages/components/ng-ui-router-layout)](https://npmjs.com/package/@ovh-ux/ng-ui-router-layout?activeTab=dependencies) | [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/manager/packages/components/ng-ui-router-layout)](https://npmjs.com/package/@ovh-ux/ng-ui-router-layout?activeTab=dependencies) | [![Peer Dependencies](https://badgen.net/david/peer/ovh-ux/manager/packages/components/ng-ui-router-layout)](https://npmjs.com/package/@ovh-ux/ng-ui-router-layout?activeTab=dependencies) | [:books:](https://github.com/ovh-ux/manager/blob/master/packages/components/ng-ui-router-layout/CHANGELOG.md) |
-| [@ovh-ux/ng-uirouter-breadcrumb](https://github.com/ovh-ux/manager/tree/master/packages/components/ng-uirouter-breadcrumb) | [![npm version](https://badgen.net/npm/v/@ovh-ux/ng-uirouter-breadcrumb)](https://www.npmjs.com/package/@ovh-ux/ng-uirouter-breadcrumb) | [![Dependencies](https://badgen.net/david/dep/ovh-ux/manager/packages/components/ng-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-uirouter-breadcrumb?activeTab=dependencies) | [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/manager/packages/components/ng-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-uirouter-breadcrumb?activeTab=dependencies) | [![Peer Dependencies](https://badgen.net/david/peer/ovh-ux/manager/packages/components/ng-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-uirouter-breadcrumb?activeTab=dependencies) | [:books:](https://github.com/ovh-ux/manager/blob/master/packages/components/ng-uirouter-breadcrumb/CHANGELOG.md) |
 | [@ovh-ux/ng-uirouter-line-progress](https://github.com/ovh-ux/manager/tree/master/packages/components/ng-uirouter-line-progress) | [![npm version](https://badgen.net/npm/v/@ovh-ux/ng-uirouter-line-progress)](https://www.npmjs.com/package/@ovh-ux/ng-uirouter-line-progress) | [![Dependencies](https://badgen.net/david/dep/ovh-ux/manager/packages/components/ng-uirouter-line-progress)](https://npmjs.com/package/@ovh-ux/ng-uirouter-line-progress?activeTab=dependencies) | [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/manager/packages/components/ng-uirouter-line-progress)](https://npmjs.com/package/@ovh-ux/ng-uirouter-line-progress?activeTab=dependencies) | [![Peer Dependencies](https://badgen.net/david/peer/ovh-ux/manager/packages/components/ng-uirouter-line-progress)](https://npmjs.com/package/@ovh-ux/ng-uirouter-line-progress?activeTab=dependencies) | [:books:](https://github.com/ovh-ux/manager/blob/master/packages/components/ng-uirouter-line-progress/CHANGELOG.md) |
 <!-- AUTO-GENERATED-CONTENT:END *-->
