+++ packages/components/ng-ui-router-breadcrumb/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ovh-ux/ng-uirouter-breadcrumb",
+  "name": "@ovh-ux/ng-ui-router-breadcrumb",
   "version": "0.1.0",
   "description": "UiRouter breadcrumb",
   "keywords": [
@@ -15,7 +15,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ovh-ux/manager.git",
-    "directory": "packages/components/ng-uirouter-breadcrumb"
+    "directory": "packages/components/ng-ui-router-breadcrumb"
   },
   "license": "BSD-3-Clause",
   "author": "OVH SAS",
@@ -24,15 +24,15 @@
   ],
   "main": "./dist/cjs/index.js",
   "module": "./dist/esm/index.js",
-  "browser": "./dist/umd/ng-uirouter-breadcrumb.js",
+  "browser": "./dist/umd/ng-ui-router-breadcrumb.js",
   "scripts": {
     "build": "rollup -c --environment BUILD:production",
     "dev": "rollup -c --environment BUILD:development",
     "dev:watch": "yarn run dev --watch",
     "prepare": "yarn run build",
-    "start": "lerna exec --stream --scope='@ovh-ux/ng-uirouter-breadcrumb' --include-filtered-dependencies -- yarn run build",
-    "start:dev": "lerna exec --stream --scope='@ovh-ux/ng-uirouter-breadcrumb' --include-filtered-dependencies -- yarn run dev",
-    "start:watch": "lerna exec --stream --parallel --scope='@ovh-ux/ng-uirouter-breadcrumb' --include-filtered-dependencies -- yarn run dev:watch"
+    "start": "lerna exec --stream --scope='@ovh-ux/ng-ui-router-breadcrumb' --include-filtered-dependencies -- yarn run build",
+    "start:dev": "lerna exec --stream --scope='@ovh-ux/ng-ui-router-breadcrumb' --include-filtered-dependencies -- yarn run dev",
+    "start:watch": "lerna exec --stream --parallel --scope='@ovh-ux/ng-ui-router-breadcrumb' --include-filtered-dependencies -- yarn run dev:watch"
   },
   "devDependencies": {
     "@ovh-ux/component-rollup-config": "^5.1.0"
