+++ packages/components/ng-ui-router-breadcrumb/src/ng-ui-router-breadcrumb.component.js
@@ -1,0 +1,7 @@
+import controller from './ng-ui-router-breadcrumb.controller';
+import template from './ng-ui-router-breadcrumb.html';
+
+export default {
+  controller,
+  template,
+};
