+++ packages/components/ng-ui-router-breadcrumb/src/ng-ui-router-breadcrumb.controller.js
@@ -1,0 +1,17 @@
+export default class BreadcrumbController {
+  /* @ngInject */
+  constructor(
+    uiRouterBreadcrumbService,
+  ) {
+    this.uiRouterBreadcrumbService = uiRouterBreadcrumbService;
+
+    this.breadcrumb = [];
+  }
+
+  $onInit() {
+    this.uiRouterBreadcrumbService.subscribe((breadcrumb) => {
+      this.breadcrumb = breadcrumb;
+    });
+    this.breadcrumb = this.uiRouterBreadcrumbService.getBreadcrumb();
+  }
+}
