+++ packages/components/ng-ui-router-breadcrumb/src/ng-ui-router-breadcrumb.less
@@ -1,4 +1,4 @@
-uirouter-breadcrumb { /* stylelint-disable-line selector-type-no-unknown */
+ui-router-breadcrumb { /* stylelint-disable-line selector-type-no-unknown */
   ul {
     margin: 0;
     padding: 0;
