+++ packages/components/ng-ui-router-breadcrumb/src/ng-ui-router-breadcrumb.service.js
