--- packages/components/ng-uirouter-breadcrumb/src/ng-uirouter-breadcrumb.component.js
@@ -1,7 +1,0 @@
-import controller from './ng-uirouter-breadcrumb.controller';
-import template from './ng-uirouter-breadcrumb.html';
-
-export default {
-  controller,
-  template,
-};
