+++ packages/manager/apps/pci/package.json
@@ -27,8 +27,8 @@
     "@ovh-ux/ng-ovh-swimming-poll": "^4.0.0",
     "@ovh-ux/ng-ovh-user-pref": "^1.0.0",
     "@ovh-ux/ng-translate-async-loader": "^2.0.0",
+    "@ovh-ux/ng-ui-router-breadcrumb": "^0.1.0",
     "@ovh-ux/ng-ui-router-layout": "^1.2.0",
-    "@ovh-ux/ng-uirouter-breadcrumb": "^0.1.0",
     "@uirouter/angularjs": "^1.0.15",
     "angular": "^1.7.5",
     "angular-animate": "^1.7.5",
