+++ packages/manager/apps/public-cloud/package.json
@@ -34,9 +34,8 @@
     "@ovh-ux/ng-ovh-swimming-poll": "^4.0.0",
     "@ovh-ux/ng-ovh-user-pref": "^1.0.0",
     "@ovh-ux/ng-translate-async-loader": "^2.0.0",
+    "@ovh-ux/ng-ui-router-breadcrumb": "^0.1.0",
     "@ovh-ux/ng-ui-router-layout": "^1.2.1",
-    "@ovh-ux/ng-uirouter-breadcrumb": "^0.1.0",
-    "@ovh-ux/ng-uirouter-layout": "^1.2.1",
     "@ovh-ux/ng-uirouter-line-progress": "^0.1.0",
     "@uirouter/angularjs": "^1.0.20",
     "angular": "^1.7.5",
