+++ packages/manager/apps/public-cloud/src/index.js
@@ -26,7 +26,7 @@ import ovhManagerPci from '@ovh-ux/manager-pci';
 import ngOvhApiWrappers from '@ovh-ux/ng-ovh-api-wrappers';
 import ngOvhOtrs from '@ovh-ux/ng-ovh-otrs';
 import ngOvhUserPref from '@ovh-ux/ng-ovh-user-pref';
-import ngUirouterBreadcrumb from '@ovh-ux/ng-uirouter-breadcrumb';
+import ngUiRouterBreadcrumb from '@ovh-ux/ng-ui-router-breadcrumb';
 import ngUirouterLineProgress from '@ovh-ux/ng-uirouter-line-progress';
 import { OTRS_URLS } from './index.constants';
 
@@ -51,7 +51,7 @@ angular
     __NG_APP_INJECTIONS__,
     uiRouter,
     atInternet,
-    ngUirouterBreadcrumb,
+    ngUiRouterBreadcrumb,
     ngUirouterLineProgress,
     ovhManagerCore,
     ovhManagerPci,
