+++ packages/manager/modules/pci/src/projects/project/project.html
@@ -28,7 +28,7 @@
                 </a>
             </div>
             <div class="pci-project-content w-100 h-100 pt-4 px-3 px-md-5" style="overflow: auto;">
-                <uirouter-breadcrumb></uirouter-breadcrumb>
+                <ui-router-breadcrumb></ui-router-breadcrumb>
                 <div class="pci-project-view mt-3 pb-4" data-ui-view>
                     <div data-ng-if="$ctrl.project">
                         <oui-page-header heading="{{ ::$ctrl.project.description }}">
